@@ -1,143 +1,187 @@
 package db
 
 import (
-	"database/sql"
 	"fmt"
 
 	"github.com/nickheyer/distroface/internal/models"
 	"golang.org/x/crypto/bcrypt"
+	"gorm.io/gorm"
 )
 
-func RunInit(db *sql.DB, cfg *models.Config) error {
+func RunInit(db *gorm.DB, cfg *models.Config) error {
 	var err error
 
 	// IF ROLES ENABLED
 	if cfg.Init.Roles {
-		if _, err = db.Exec(rolesSQL); err != nil {
+		if err = createDefaultRoles(db); err != nil {
 			return fmt.Errorf("failed to init roles: %w", err)
 		}
 	}
 
 	// IF GROUPS ENABLED
 	if cfg.Init.Groups {
-		if _, err = db.Exec(groupsSQL); err != nil {
+		if err = createDefaultGroups(db); err != nil {
 			return fmt.Errorf("failed to init groups: %w", err)
 		}
 	}
 
 	// IF USER ENABLED
 	if cfg.Init.User {
-		userSQL, err := genUserSQL(cfg)
-		if err != nil {
-			return fmt.Errorf("failed generating user insert SQL: %w", err)
-		}
-
-		if _, err := db.Exec(userSQL); err != nil {
+		if err = createAdminUser(db, cfg); err != nil {
 			return fmt.Errorf("failed to init user: %w", err)
 		}
 	}
+
 	return nil
 }
 
-const rolesSQL = `
-INSERT OR IGNORE INTO roles (name, description, permissions) VALUES 
-('anonymous', 'Unauthenticated access', 
- '[
-    {"action":"VIEW","resource":"WEBUI"},
-    {"action":"LOGIN","resource":"WEBUI"},
-    {"action":"PULL","resource":"IMAGE"},
-    {"action":"VIEW","resource":"TAG"},
-    {"action":"VIEW","resource":"ARTIFACT"},
-    {"action":"DOWNLOAD","resource":"ARTIFACT"},
-    {"action":"VIEW","resource":"REPO"}
- ]'),
+func createDefaultRoles(db *gorm.DB) error {
+	roles := []models.Role{
+		{
+			Name:        "anonymous",
+			Description: "Unauthenticated access",
+			Permissions: models.PermissionArray{
+				{Action: models.ActionView, Resource: models.ResourceWebUI},
+				{Action: models.ActionLogin, Resource: models.ResourceWebUI},
+				{Action: models.ActionPull, Resource: models.ResourceImage},
+				{Action: models.ActionView, Resource: models.ResourceTag},
+				{Action: models.ActionView, Resource: models.ResourceArtifact},
+				{Action: models.ActionDownload, Resource: models.ResourceArtifact},
+				{Action: models.ActionView, Resource: models.ResourceRepo},
+			},
+		},
+		{
+			Name:        "reader",
+			Description: "Basic read access",
+			Permissions: models.PermissionArray{
+				{Action: models.ActionView, Resource: models.ResourceWebUI},
+				{Action: models.ActionLogin, Resource: models.ResourceWebUI},
+				{Action: models.ActionPull, Resource: models.ResourceImage},
+				{Action: models.ActionView, Resource: models.ResourceTag},
+				{Action: models.ActionView, Resource: models.ResourceUser},
+				{Action: models.ActionView, Resource: models.ResourceGroup},
+				{Action: models.ActionView, Resource: models.ResourceArtifact},
+				{Action: models.ActionDownload, Resource: models.ResourceArtifact},
+				{Action: models.ActionView, Resource: models.ResourceRepo},
+			},
+		},
+		{
+			Name:        "developer",
+			Description: "Standard developer access",
+			Permissions: models.PermissionArray{
+				{Action: models.ActionMigrate, Resource: models.ResourceTask},
+				{Action: models.ActionView, Resource: models.ResourceWebUI},
+				{Action: models.ActionLogin, Resource: models.ResourceWebUI},
+				{Action: models.ActionUpdate, Resource: models.ResourceWebUI},
+				{Action: models.ActionPull, Resource: models.ResourceImage},
+				{Action: models.ActionPush, Resource: models.ResourceImage},
+				{Action: models.ActionView, Resource: models.ResourceImage},
+				{Action: models.ActionView, Resource: models.ResourceTag},
+				{Action: models.ActionCreate, Resource: models.ResourceTag},
+				{Action: models.ActionDelete, Resource: models.ResourceTag},
+				{Action: models.ActionView, Resource: models.ResourceUser},
+				{Action: models.ActionView, Resource: models.ResourceGroup},
+				{Action: models.ActionView, Resource: models.ResourceArtifact},
+				{Action: models.ActionUpload, Resource: models.ResourceArtifact},
+				{Action: models.ActionUpdate, Resource: models.ResourceArtifact},
+				{Action: models.ActionDownload, Resource: models.ResourceArtifact},
+				{Action: models.ActionDelete, Resource: models.ResourceArtifact},
+				{Action: models.ActionView, Resource: models.ResourceRepo},
+				{Action: models.ActionCreate, Resource: models.ResourceRepo},
+				{Action: models.ActionDelete, Resource: models.ResourceRepo},
+			},
+		},
+		{
+			Name:        "administrator",
+			Description: "Full system access",
+			Permissions: models.PermissionArray{
+				{Action: models.ActionAdmin, Resource: models.ResourceSystem},
+				{Action: models.ActionView, Resource: models.ResourceWebUI},
+				{Action: models.ActionLogin, Resource: models.ResourceWebUI},
+				{Action: models.ActionView, Resource: models.ResourceUser},
+				{Action: models.ActionCreate, Resource: models.ResourceUser},
+				{Action: models.ActionUpdate, Resource: models.ResourceUser},
+				{Action: models.ActionDelete, Resource: models.ResourceUser},
+				{Action: models.ActionView, Resource: models.ResourceGroup},
+				{Action: models.ActionCreate, Resource: models.ResourceGroup},
+				{Action: models.ActionUpdate, Resource: models.ResourceGroup},
+				{Action: models.ActionDelete, Resource: models.ResourceGroup},
+				{Action: models.ActionUpdate, Resource: models.ResourceWebUI},
+				{Action: models.ActionPull, Resource: models.ResourceImage},
+				{Action: models.ActionPush, Resource: models.ResourceImage},
+				{Action: models.ActionMigrate, Resource: models.ResourceTask},
+				{Action: models.ActionDelete, Resource: models.ResourceImage},
+				{Action: models.ActionView, Resource: models.ResourceTag},
+				{Action: models.ActionCreate, Resource: models.ResourceTag},
+				{Action: models.ActionDelete, Resource: models.ResourceTag},
+				{Action: models.ActionView, Resource: models.ResourceArtifact},
+				{Action: models.ActionUpload, Resource: models.ResourceArtifact},
+				{Action: models.ActionUpdate, Resource: models.ResourceArtifact},
+				{Action: models.ActionDownload, Resource: models.ResourceArtifact},
+				{Action: models.ActionDelete, Resource: models.ResourceArtifact},
+				{Action: models.ActionView, Resource: models.ResourceRepo},
+				{Action: models.ActionCreate, Resource: models.ResourceRepo},
+				{Action: models.ActionDelete, Resource: models.ResourceRepo},
+			},
+		},
+	}
 
-('reader', 'Basic read access', 
- '[
-    {"action":"VIEW","resource":"WEBUI"},
-    {"action":"LOGIN","resource":"WEBUI"},
-    {"action":"PULL","resource":"IMAGE"},
-    {"action":"VIEW","resource":"TAG"},
-    {"action":"VIEW","resource":"USER"},
-    {"action":"VIEW","resource":"GROUP"},
-    {"action":"VIEW","resource":"ARTIFACT"},
-    {"action":"DOWNLOAD","resource":"ARTIFACT"},
-    {"action":"VIEW","resource":"REPO"}
- ]'),
+	for _, role := range roles {
+		result := db.Where("name = ?", role.Name).FirstOrCreate(&role)
+		if result.Error != nil {
+			return result.Error
+		}
+	}
 
-('developer', 'Standard developer access', 
- '[
-    {"action":"MIGRATE","resource":"TASK"},
-    {"action":"VIEW","resource":"WEBUI"},
-    {"action":"LOGIN","resource":"WEBUI"},
-    {"action":"UPDATE","resource":"WEBUI"},
-    {"action":"PULL","resource":"IMAGE"},
-    {"action":"PUSH","resource":"IMAGE"},
-    {"action":"VIEW","resource":"IMAGE"},
-    {"action":"VIEW","resource":"TAG"},
-    {"action":"CREATE","resource":"TAG"},
-    {"action":"DELETE","resource":"TAG"},
-    {"action":"VIEW","resource":"USER"},
-    {"action":"VIEW","resource":"GROUP"},
-    {"action":"VIEW","resource":"ARTIFACT"},
-    {"action":"UPLOAD","resource":"ARTIFACT"},
-    {"action":"UPDATE","resource":"ARTIFACT"},
-    {"action":"DOWNLOAD","resource":"ARTIFACT"},
-    {"action":"DELETE","resource":"ARTIFACT"},
-    {"action":"VIEW","resource":"REPO"},
-    {"action":"CREATE","resource":"REPO"},
-    {"action":"DELETE","resource":"REPO"}
- ]'),
+	return nil
+}
 
-('administrator', 'Full system access', 
- '[
-    {"action":"ADMIN","resource":"SYSTEM"},
-    {"action":"VIEW","resource":"WEBUI"},
-    {"action":"LOGIN","resource":"WEBUI"},
-    {"action":"VIEW","resource":"USER"},
-    {"action":"CREATE","resource":"USER"},
-    {"action":"UPDATE","resource":"USER"},
-    {"action":"DELETE","resource":"USER"},
-    {"action":"VIEW","resource":"GROUP"},
-    {"action":"CREATE","resource":"GROUP"},
-    {"action":"UPDATE","resource":"GROUP"},
-    {"action":"DELETE","resource":"GROUP"},
-    {"action":"UPDATE","resource":"WEBUI"},
-    {"action":"PULL","resource":"IMAGE"},
-    {"action":"PUSH","resource":"IMAGE"},
-    {"action":"MIGRATE","resource":"TASK"},
-    {"action":"DELETE","resource":"IMAGE"},
-    {"action":"VIEW","resource":"TAG"},
-    {"action":"CREATE","resource":"TAG"},
-    {"action":"DELETE","resource":"TAG"},
-    {"action":"VIEW","resource":"ARTIFACT"},
-    {"action":"UPLOAD","resource":"ARTIFACT"},
-    {"action":"UPDATE","resource":"ARTIFACT"},
-    {"action":"DOWNLOAD","resource":"ARTIFACT"},
-    {"action":"DELETE","resource":"ARTIFACT"},
-    {"action":"VIEW","resource":"REPO"},
-    {"action":"CREATE","resource":"REPO"},
-    {"action":"DELETE","resource":"REPO"}
- ]');
-`
+func createDefaultGroups(db *gorm.DB) error {
+	groups := []models.Group{
+		{
+			Name:        "admins",
+			Description: "System Administrators",
+			Roles:       models.StringArray{"administrator"},
+			Scope:       "system:all",
+		},
+		{
+			Name:        "developers",
+			Description: "Development Team",
+			Roles:       models.StringArray{"developer"},
+			Scope:       "system:all",
+		},
+		{
+			Name:        "readers",
+			Description: "Read-only Users",
+			Roles:       models.StringArray{"reader"},
+			Scope:       "system:all",
+		},
+	}
 
-const groupsSQL = `
-INSERT OR IGNORE INTO groups (name, description, roles, scope) VALUES 
-('admins', 'System Administrators', '["administrator"]', 'system:all'),
-('developers', 'Development Team', '["developer"]', 'system:all'),
-('readers', 'Read-only Users', '["reader"]', 'system:all');
-`
+	for _, group := range groups {
+		result := db.Where("name = ?", group.Name).FirstOrCreate(&group)
+		if result.Error != nil {
+			return result.Error
+		}
+	}
 
-func genUserSQL(cfg *models.Config) (string, error) {
+	return nil
+}
+
+func createAdminUser(db *gorm.DB, cfg *models.Config) error {
 	// GENERATE BCRYPT HASH FOR PASS
 	hash, err := bcrypt.GenerateFromPassword([]byte(cfg.Init.Password), bcrypt.DefaultCost)
 	if err != nil {
-		return "", fmt.Errorf("failed to hash password for user %q: %w", cfg.Init.Username, err)
+		return fmt.Errorf("failed to hash password for user %q: %w", cfg.Init.Username, err)
 	}
 
-	// RETURN GENERATED SQL, PUT USER IN ADMINS FOR EZ
-	return fmt.Sprintf(`
-INSERT OR IGNORE INTO users (username, password, groups) 
-VALUES ('%s', '%s', '["admins"]');
-`, cfg.Init.Username, hash), nil
+	// PUT USER IN ADMINS FOR EZ
+	user := models.User{
+		Username: cfg.Init.Username,
+		Password: hash,
+		Groups:   models.StringArray{"admins"},
+	}
+
+	result := db.Where("username = ?", user.Username).FirstOrCreate(&user)
+	return result.Error
 }
